+++ manifest.config.ts
@@ -7,7 +7,7 @@ export default defineManifest((env: ConfigEnv) => {
   const isDev = env.mode === 'development'
 
   return {
-    name: "MetaDock - Builders' Swiss Army Knife",
+    name: "MetaSuites - Builders' Swiss Army Knife",
     description:
       'Integrating 30+ popular Web3 and AI Tools, 300M+ Address Labels, and the Fund Flow Map into a single extension.',
     version: version,
@@ -37,7 +37,8 @@ export default defineManifest((env: ConfigEnv) => {
               '*://*.wemixscan.com/*',
               '*://gnosisscan.io/*',
               '*://*.celoscan.io/*',
-              '*://*.bttcscan.com/*'
+              '*://*.bttcscan.com/*',
+              '*://era.zksync.network/*'
             ]
           : ['<all_urls>'],
         js: ['src/content/index.all_frames.ts'],