+++ src/content/scans/components/Button/index.tsx
@@ -55,7 +55,6 @@ const Button: FC<PropsWithChildren<Props>> = props => {
           href={href}
           style={stylesheet}
           target="_blank"
-          rel="noopener noreferrer"
           onClick={onClick}
         >
           {icon && icon}