+++ src/pages/PrivacyPolicy/App.tsx
@@ -12,10 +12,10 @@ const App: FC = () => {
       </header>
       <div className={styles.content}>
         <div>
-          At MetaDock, accessible from blocksec.com, one of our main priorities
-          is the privacy of our visitors. This Privacy Policy document contains
-          types of information that is collected and recorded by MetaDock and
-          how we use it.
+          At MetaSuites, accessible from blocksec.com, one of our main
+          priorities is the privacy of our visitors. This Privacy Policy
+          document contains types of information that is collected and recorded
+          by MetaSuites and how we use it.
           <br />
           <br />
           If you have additional questions or require more information about our
@@ -24,7 +24,7 @@ const App: FC = () => {
           <br />
           This Privacy Policy applies only to our online activities and is valid
           for visitors to our website with regards to the information that they
-          shared and/or collect in MetaDock. This policy is not applicable to
+          shared and/or collect in MetaSuites. This policy is not applicable to
           any information collected offline or via channels other than this
           website. Our Privacy Policy was created with the help of the Free
           Privacy Policy Generator.
@@ -73,9 +73,9 @@ const App: FC = () => {
         </div>
         <h2 className={styles.title}>Log Files</h2>
         <div>
-          MetaDock follows a standard procedure of using log files. These files
-          log visitors when they visit websites. All hosting companies do this
-          and a part of hosting services&apos; analytics. The information
+          MetaSuites follows a standard procedure of using log files. These
+          files log visitors when they visit websites. All hosting companies do
+          this and a part of hosting services&apos; analytics. The information
           collected by log files include internet protocol (IP) addresses,
           browser type, Internet Service Provider (ISP), date and time stamp,
           referring/exit pages, and possibly the number of clicks. These are not
@@ -87,26 +87,26 @@ const App: FC = () => {
         <h2 className={styles.title}>Advertising Partners Privacy Policies</h2>
         <div>
           You may consult this list to find the Privacy Policy for each of the
-          advertising partners of MetaDock.
+          advertising partners of MetaSuites.
           <br />
           <br />
           Third-party ad servers or ad networks uses technologies like cookies,
           JavaScript, or Web Beacons that are used in their respective
-          advertisements and links that appear on MetaDock, which are sent
+          advertisements and links that appear on MetaSuites, which are sent
           directly to users&apos; browser. They automatically receive your IP
           address when this occurs. These technologies are used to measure the
           effectiveness of their advertising campaigns and/or to personalize the
           advertising content that you see on websites that you visit.
           <br />
           <br />
-          Note that MetaDock has no access to or control over these cookies that
-          are used by third-party advertisers.
+          Note that MetaSuites has no access to or control over these cookies
+          that are used by third-party advertisers.
         </div>
         <h2 className={styles.title}>Third Party Privacy Policies</h2>
         <div>
-          MetaDock&apos;s Privacy Policy does not apply to other advertisers or
-          websites. Thus, we are advising you to consult the respective Privacy
-          Policies of these third-party ad servers for more detailed
+          MetaSuites&apos;s Privacy Policy does not apply to other advertisers
+          or websites. Thus, we are advising you to consult the respective
+          Privacy Policies of these third-party ad servers for more detailed
           information. It may include their practices and instructions about how
           to opt-out of certain options.
           <br />
@@ -184,7 +184,7 @@ const App: FC = () => {
           participate in, and/or monitor and guide their online activity.
           <br />
           <br />
-          MetaDock does not knowingly collect any Personal Identifiable
+          MetaSuites does not knowingly collect any Personal Identifiable
           Information from children under the age of 13. If you think that your
           child provided this kind of information on our website, we strongly
           encourage you to contact us immediately and we will do our best